pyproject.toml
[project]
name = "mcp-managed-agents"
version = "0.1.0"
description = "MCP server wrapping Anthropic's Managed Agents API"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
]

[project.scripts]
mcp-managed-agents = "mcp_managed_agents.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
